Refactor commonly used functions
